requirements.txt
ultralytics~=8.2.21
pip~=23.2.1
pillow~=10.3.0
Jinja2~=3.1.4
typing_extensions~=4.12.0
packaging~=24.0
Augmentor~=0.2.12
opencv-python~=4.9.0.80
Albumentations